backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.1